refactor(db,scripts): make dedicated /scripts package, move DB scripts
