+++ libs/db/README.md
@@ -1,4 +1,3 @@
 # db
 
-This library stores our Prisma schema and migrations (which are sacred to us),
-as well as db-related utility scripts used in development and testing.
+Home of the Prisma schema and migrations, and any other SQL-related code.
