+++ libs/db/src/index.ts
@@ -1,2 +1,2 @@
-export * from './prisma-utils/utils';
-export * from './prisma-utils/extended-client';
+export * from './utils';
+export * from './extended-client';
