+++ libs/db/tsconfig.json
@@ -2,12 +2,12 @@
   "extends": "../../tsconfig.base.json",
   "files": [],
   "include": [],
+  "compilerOptions": {
+    "module": "commonjs"
+  },
   "references": [
     {
       "path": "./tsconfig.lib.json"
-    },
-    {
-      "path": "./tsconfig.scripts.json"
     }
   ]
 }
