+++ scripts/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@momentum/scripts",
+  "type": "commonjs"
+}
