+++ scripts/src/reset.script.ts
@@ -1,0 +1,5 @@
+import { nuke } from '@momentum/db';
+
+import { prismaWrapper } from './prisma-wrapper.util';
+
+prismaWrapper(async (prisma) => await nuke(prisma));
